refactor: replace PhosphorIcon with AppIcon in exports
